requirements.txt
pymupdf==1.24.10
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy>=1.24
tqdm>=4.66
pydantic>=2.7
python-dotenv>=1.0
requests>=2.31
regex>=2024.5.15
